refactor: move test logger
