+++ utils/logging/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
         "level.go",
         "log.go",
         "logger.go",
+        "no_log.go",
         "sanitize.go",
-        "test_log.go",
     ],
     importpath = "github.com/ava-labs/avalanchego/utils/logging",
     visibility = ["//visibility:public"],
