+++ utils/logging/loggingtest/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "loggingtest",
+    testonly = True,
+    srcs = ["logging.go"],
+    importpath = "github.com/ava-labs/avalanchego/utils/logging/loggingtest",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//utils/logging",
+        "@org_uber_go_zap//:zap",
+        "@org_uber_go_zap//zapcore",
+    ],
+)
