+++ vms/saevm/blocks/BUILD.bazel
@@ -55,6 +55,7 @@ go_test(
     embed = [":blocks"],
     deps = [
         "//utils/logging",
+        "//utils/logging/loggingtest",
         "//vms/components/gas",
         "//vms/saevm/cmputils",
         "//vms/saevm/gastime",
