+++ vms/saevm/blocks/blockstest/BUILD.bazel
@@ -12,11 +12,11 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//utils/logging",
+        "//utils/logging/loggingtest",
         "//vms/components/gas",
         "//vms/saevm/blocks",
         "//vms/saevm/hook",
         "//vms/saevm/hook/hookstest",
-        "//vms/saevm/saetest",
         "//vms/saevm/types",
         "@com_github_ava_labs_libevm//common",
         "@com_github_ava_labs_libevm//core",
