+++ vms/saevm/blocks/execution_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/ava-labs/avalanchego/utils/logging"
+	"github.com/ava-labs/avalanchego/utils/logging/loggingtest"
 	"github.com/ava-labs/avalanchego/vms/components/gas"
 	"github.com/ava-labs/avalanchego/vms/saevm/cmputils"
 	"github.com/ava-labs/avalanchego/vms/saevm/gastime"
@@ -141,7 +142,7 @@ func TestMarkExecuted(t *testing.T) {
 			}
 
 			t.Run("MarkExecuted_again", func(t *testing.T) {
-				rec := saetest.NewLogRecorder(logging.Warn)
+				rec := loggingtest.NewLogRecorder(logging.Warn)
 				b.log = rec
 				require.ErrorIs(t, b.MarkExecuted(db, xdb, gasTime, wallTime, baseFee.ToBig(), receipts, stateRoot, lastExecuted), errMarkBlockExecutedAgain)
 				// The database's head block might have been corrupted so this MUST
