+++ vms/saevm/cchain/state/BUILD.bazel
@@ -52,10 +52,10 @@ go_test(
         "//ids",
         "//snow/snowtest",
         "//utils/logging",
+        "//utils/logging/loggingtest",
         "//vms/components/avax",
         "//vms/saevm/cchain/tx",
         "//vms/saevm/cchain/tx/txtest",
-        "//vms/saevm/saetest",
         "//vms/secp256k1fx",
         "@com_github_ava_labs_libevm//common",
         "@com_github_ava_labs_libevm//core/types",
