+++ vms/saevm/sae/BUILD.bazel
@@ -95,6 +95,7 @@ go_test(
         "//snow/validators/validatorstest",
         "//utils",
         "//utils/logging",
+        "//utils/logging/loggingtest",
         "//utils/set",
         "//utils/timer/mockable",
         "//version",
@@ -180,6 +181,7 @@ go_test(
         "//snow/validators/validatorstest",
         "//utils",
         "//utils/logging",
+        "//utils/logging/loggingtest",
         "//utils/set",
         "//utils/timer/mockable",
         "//version",
