+++ vms/saevm/saetest/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
     testonly = True,
     srcs = [
         "heightdb.go",
-        "logging.go",
         "saetest.go",
         "wallet.go",
     ],
@@ -15,7 +14,6 @@ go_library(
     deps = [
         "//database",
         "//utils/lock",
-        "//utils/logging",
         "//vms/saevm/saedb",
         "//vms/saevm/types",
         "@com_github_ava_labs_libevm//common",
@@ -29,8 +27,6 @@ go_library(
         "@com_github_ava_labs_libevm//trie",
         "@com_github_holiman_uint256//:uint256",
         "@com_github_stretchr_testify//require",
-        "@org_uber_go_zap//:zap",
-        "@org_uber_go_zap//zapcore",
     ],
 )
